+++ Source/Log/Log.cpp
@@ -19,17 +19,21 @@
 // Qt5
 #include <QtCore/QDebug>
 #include <QtCore/QDateTime>
+#ifdef QT_GUI_LIB
+#include <QtWidgets/QListWidget>
+#endif
 // C++ STL
 #include <iostream>
 
 #define CUR_DT_STR QDateTime::currentDateTimeUtc().toString(Qt::ISODateWithMs)
 
 #define RED "\033[031m"
-#define YELLOW "\033[033m"
-#define MAGENTA "\033[035m"
 #define GREEN "\033[032m"
-#define CYAN "\033[036m"
+#define YELLOW "\033[033m"
 #define BLUE "\033[034m"
+#define MAGENTA "\033[035m"
+#define CYAN "\033[036m"
+#define GRAY "\033[037m"
 #define NC "\033[0m"
 
 namespace Ramio {
@@ -38,6 +42,10 @@ QString timeLogFormatStr()
 {
 	return CUR_DT_STR;
 }
+
+Log::Log() = default;
+
+Log::~Log() = default;
 
 const QString& Log::log(const QString& str)
 {
@@ -51,11 +59,21 @@ const QString& Log::ulog(const QString& str)
 	userLog_.append(LogRecord{QDateTime::currentDateTime(), str, 0});
 	printLogRecord(userLog_.last());
 #endif
-	qInfo().noquote().nospace()<<CUR_DT_STR<<CYAN<<" [info] "<<NC<<str;
+	qInfo().noquote().nospace()<<CUR_DT_STR<<GRAY<<" [info] "<<NC<<str;
 	return str;
 }
 
-const QString &Log::wlog(const QString &str)
+const QString& Log::nlog(const QString& title, const QString& text)
+{
+#ifdef QT_GUI_LIB
+	if (noticer_)
+		noticer_->notice(QDateTime::currentDateTime(), title, text);
+#endif
+	qInfo().noquote().nospace()<<CUR_DT_STR<<CYAN<<" [notice] "<<NC << title << ": " <<text;
+	return text;
+}
+
+const QString& Log::wlog(const QString &str)
 {
 #ifdef QT_GUI_LIB
 	userLog_.append(LogRecord{QDateTime::currentDateTime(), str, 1});
@@ -101,6 +119,11 @@ void Log::setLogWidget(QListWidget* widget)
 	if (logWidget_)
 		for (const LogRecord& record: userLog_)
 			printLogRecord(record);
+}
+
+void Log::setNoticer(Noticer* noticer)
+{
+	noticer_ = noticer;
 }
 
 void Log::clearHistoryUWLog()